Separate game endpoint
